backend/requirements.txt
APScheduler==3.6.3
astroid==2.4.2
autopep8==1.5.4
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.3
click==7.1.2
decorator==4.4.2
dnspython==1.16.0
email-validator==1.1.1
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
ipython==7.15.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.1
parso==0.7.0
passlib==1.7.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
pylint==2.5.3
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.17
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
tzlocal==2.1
wcwidth==0.2.4
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
